requirements.txt
Flask==2.3.3
PyCryptodome==3.19.0
Werkzeug==2.3.7
python-dotenv==1.0.0



